+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `CAlloc::allocation_usable_size`
 - `{Global,}Tlsf::new`
 - `FlexTlsf::new` as a `const fn`
+- `ConstDefault` implementation for `Tlsf`
 
 ### Removed